These are some small projects | made using JavaScript
These are some small projects | made using JavaScript

Projects

* Pizza Menu
* Tic-Tac-Toe Game
* Calculator

Pizza Menu
This project is a menu used to order pizza with a various selection of toppings. It then takes the input information and calculates a price for the order Tic-Tac-Toe Game

The classic game of tic-tac-toe. It randomly rolls dice to select which of the two players goes first and then users are off playing the classic game. Users can decide which token each player wants to use and the winner is celebrated with music and flashing lights.

Calculator

A basic 4 function calculator. Users can add, subtract, multiply, and divide on a sleek web version of a calculator.